Fix Typo
There was a typo in the annotation of the simple repeats. This lead to problem that the simple repeat annotation was not used in the prioritization step. This error only occurred when the annotation was initiated through AIdiva, otherwise if AIdiva was run on already annotated data there shouldn't be any problem.