Complete RESTRUCTURE of the repository.
The annotation part of the tool can be used to get from an VCF file to the required input table, but is not mandatory.

aiDIVA itself requires as input a TAB separated table with the all the features present. This also makes it easier to integrate it into existing pipelines.

- ADDED clear run scripts for the different modi of aiDIVA (aiDIVA-RF, aiDIVA-meta, aiDIVA-meta-RF)
- ADDED run script for separate annotation part
- ADDED separate example configuration files for aiDIVA and annotation (ATTENTION: they differ in what they provide so make sure to use the correct one!)
- MOVED constants from the scripts to the configuration file (makes it easier if they need to be changed)
- FIXED Fixed pandas future warnings -> to prevent problems with future versions